backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
pydantic
pydantic-settings
aiofiles
cors